pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forge"
version = "0.0.1"
description = "Model-agnostic, self-learning, self-healing agent harness"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Justin Bell" }]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.6",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
all = ["anthropic>=0.40.0", "openai>=1.50.0"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
forge = "forge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["forge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
